integrate notification feature into the website
