+++ app/views/schedule.html.erb
@@ -200,9 +200,6 @@
 </div>
 
 
-
-
-
 </body>
 </html>
